split up project resource specs
